Revert "Export NLF"
This reverts commit 4e3cdf265b259d4af58f92c754c2d906489e02a7.

+++ nlf/pt/multiperson/person_detector.py
@@ -158,7 +158,7 @@ class PersonDetector(nn.Module):
         return boxes, scores
 
     def call_model(self, images):
-        preds = self.model(images.to(dtype=torch.float16)).float()
+        preds = self.model(images.to(dtype=torch.float16))
         preds = torch.permute(preds, [0, 2, 1])  # [batch, n_boxes, 84]
         boxes = preds[..., :4]
         scores = preds[..., 4:]
